Issue #142 refactor: 스크롤 인덱스 구하기, 오늘의 스케줄 구하기 viewModel로 이동
